(PC-36379)[PRO] fix: fix stock display condition in IndividualOfferPriceTable
