+++ pro/src/pages/IndividualOffer/IndividualOfferPriceTable/components/IndividualOfferPriceTableScreen.tsx
@@ -1,13 +1,11 @@
 import { yupResolver } from '@hookform/resolvers/yup'
 import { FormProvider, useForm } from 'react-hook-form'
 import { useLocation, useNavigate } from 'react-router'
-import { useSWRConfig } from 'swr'
 
 import type {
   GetIndividualOfferWithAddressResponseModel,
   GetOfferStockResponseModel,
 } from '@/apiClient/v1'
-import { GET_OFFER_QUERY_KEY } from '@/commons/config/swrQueryKeys'
 import { useIndividualOfferContext } from '@/commons/context/IndividualOfferContext/IndividualOfferContext'
 import { REIMBURSEMENT_RULES } from '@/commons/core/Finances/constants'
 import {
@@ -19,21 +17,15 @@ import { isOfferDisabled as getIsOfferDisabled } from '@/commons/core/Offers/uti
 import { assertOrFrontendError } from '@/commons/errors/assertOrFrontendError'
 import { useActiveFeature } from '@/commons/hooks/useActiveFeature'
 import { useIsCaledonian } from '@/commons/hooks/useIsCaledonian'
-import { useNotification } from '@/commons/hooks/useNotification'
 import { useOfferWizardMode } from '@/commons/hooks/useOfferWizardMode'
 import { DuoCheckbox } from '@/components/DuoCheckbox/DuoCheckbox'
 import { FormLayout } from '@/components/FormLayout/FormLayout'
-import { getSuccessMessage } from '@/components/IndividualOffer/utils/getSuccessMessage'
 import { RouteLeavingGuardIndividualOffer } from '@/components/RouteLeavingGuardIndividualOffer/RouteLeavingGuardIndividualOffer'
 import { ActionBar } from '@/pages/IndividualOffer/components/ActionBar/ActionBar'
 
-import {
-  type PriceTableFormValues,
-  PriceTableValidationSchema,
-} from '../commons/schemas'
+import { useSaveOfferPriceTable } from '../commons/hooks/useSaveOfferPriceTable'
+import { PriceTableValidationSchema } from '../commons/schemas'
 import type { PriceTableFormContext } from '../commons/types'
-import { saveEventOfferPriceTable } from '../commons/utils/saveEventOfferPriceTable'
-import { saveNonEventOfferPriceTable } from '../commons/utils/saveNonEventOfferPriceTable'
 import { toFormValues } from '../commons/utils/toFormValues'
 import { ActivationCodeCallout } from './ActivationCodeCallout'
 import { NonRefundableCallout } from './NonRefundableCallout'
@@ -46,14 +38,12 @@ interface IndividualOfferPriceTableScreenProps {
 export const IndividualOfferPriceTableScreen = ({
   offer,
   offerStocks,
-}: IndividualOfferPriceTableScreenProps): JSX.Element => {
+}: IndividualOfferPriceTableScreenProps) => {
   const mode = useOfferWizardMode()
   const navigate = useNavigate()
   const { pathname } = useLocation()
-  const notify = useNotification()
   const { subCategories, hasPublishedOfferWithSameEan } =
     useIndividualOfferContext()
-  const { mutate } = useSWRConfig()
   const isMediaPageEnabled = useActiveFeature('WIP_ADD_VIDEO')
   const isCaledonian = useIsCaledonian()
 
@@ -85,53 +75,10 @@ export const IndividualOfferPriceTableScreen = ({
     resolver: yupResolver(PriceTableValidationSchema),
   })
 
-  // TODO (igabriele, 2025-08-22): Move that into a hook.
-  const onSubmit = async (formValues: PriceTableFormValues): Promise<void> => {
-    const nextStepUrl = getIndividualOfferUrl({
-      offerId: offer.id,
-      step:
-        mode === OFFER_WIZARD_MODE.EDITION
-          ? INDIVIDUAL_OFFER_WIZARD_STEP_IDS.TARIFS
-          : offer.isEvent
-            ? INDIVIDUAL_OFFER_WIZARD_STEP_IDS.STOCKS
-            : INDIVIDUAL_OFFER_WIZARD_STEP_IDS.SUMMARY,
-      mode:
-        mode === OFFER_WIZARD_MODE.EDITION ? OFFER_WIZARD_MODE.READ_ONLY : mode,
-      isOnboarding,
-    })
-
-    if (!form.formState.isDirty && mode === OFFER_WIZARD_MODE.EDITION) {
-      navigate(nextStepUrl)
-
-      notify.success(getSuccessMessage(mode))
-
-      return
-    }
-
-    try {
-      if (offer.isEvent) {
-        await saveEventOfferPriceTable(formValues, { offer })
-      } else {
-        await saveNonEventOfferPriceTable(formValues, {
-          offer,
-          setError: form.setError,
-        })
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        notify.error(error.message)
-      }
-      return
-    }
-
-    await mutate([GET_OFFER_QUERY_KEY, offer.id])
-
-    if (mode === OFFER_WIZARD_MODE.EDITION) {
-      notify.success(getSuccessMessage(mode))
-    }
-
-    navigate(nextStepUrl)
-  }
+  const { saveAndContinue } = useSaveOfferPriceTable({
+    form,
+    offer,
+  })
 
   const handlePreviousStepOrBackToReadOnly = () => {
     if (mode === OFFER_WIZARD_MODE.EDITION) {
@@ -161,7 +108,7 @@ export const IndividualOfferPriceTableScreen = ({
     <>
       <FormProvider {...form}>
         <form
-          onSubmit={form.handleSubmit(onSubmit)}
+          onSubmit={form.handleSubmit(saveAndContinue)}
           data-testid="stock-thing-form"
         >
           <FormLayout>
